# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar AI Workbench API
  version: 0.1.0
  description: |
    模块化桌面 AI 工作台后端契约（Stellar AI Workbench, 作者：晨星）。
    所有端点前缀 /api/v1。对话/智能体走 SSE 流式。
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8000
    description: 本地后端（Tauri sidecar 内运行）
...
paths:
...
  /api/v1/knowledge/add:
    post:
      tags: [knowledge]
      summary: 导入文本到知识库（切片 + 向量化落盘）
      operationId: addKnowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kb_id, texts]
              properties:
                kb_id: { type: string, example: demo }
                texts:
                  type: array
                  items: { type: string }
                metas:
                  type: array
                  items: { type: object }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  kb_id: { type: string }
                  chunk_count: { type: integer, example: 1 }
  /api/v1/rag/query:
    post:
      tags: [rag]
      summary: 检索增强问答（返回答案 + 来源片段）
      operationId: ragQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, kb_id]
              properties:
                query: { type: string, example: Stellar 是什么 }
                kb_id: { type: string, example: demo }
                top_k: { type: integer, default: 4 }
                model: { type: string, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        text: { type: string }
                        score: { type: number }
                        meta: { type: object }
...